Move the run command
